Move person admin classes to person app
