+++ ynr/apps/people/admin.py
@@ -1,5 +1,11 @@
 from django.contrib import admin
+from django.forms import ModelForm
 
+from candidates.models import (
+    ComplexPopoloField,
+    PersonExtraFieldValue,
+    ExtraField,
+)
 from popolo.models import Membership
 from people.models import Person
 
@@ -49,6 +55,30 @@ class PersonAdmin(admin.ModelAdmin):
         ),
     )
     # inlines = generics.BASE_INLINES + [MembershipInline]
+
+
+class PersonExtraFieldValueAdminForm(ModelForm):
+    pass
+
+
+class ExtraFieldAdminForm(ModelForm):
+    pass
+
+
+@admin.register(ComplexPopoloField)
+class ComplexPopoloFieldAdmin(admin.ModelAdmin):
+    list_display = ["name", "label", "order"]
+    ordering = ("order",)
+
+
+@admin.register(PersonExtraFieldValue)
+class PersonExtraFieldValueAdmin(admin.ModelAdmin):
+    form = PersonExtraFieldValueAdminForm
+
+
+@admin.register(ExtraField)
+class ExtraFieldAdmin(admin.ModelAdmin):
+    form = ExtraFieldAdminForm
 
 
 admin.site.register(Person, PersonAdmin)
